FEATURE: Allow version numbers in full namespace
Migration classes are usually named like "Version20200911105700" to be
both, naturally sorted and with little chance to conflict.

This patch allows for migration classes with arbitrary names but
living in namespaces that contain the version number part.

This patch does not change how migratins are detected within the file
system.